+++ core/src/main/java/arez/ObservableValue.java
@@ -402,6 +402,10 @@ public final class ObservableValue<T>
                        "named '" + getName() + "' where the observer is scheduled at a " + observer.getPriority() +
                        " priority but the ObservableValue's owner is scheduled at a " +
                        getOwner().getPriority() + " priority." );
+      invariant( () -> getContext().getTransaction().getTracker() == observer,
+                 () -> "Arez-0203: Attempting to add observer named '" + observer.getName() + "' to ObservableValue " +
+                       "named '" + getName() + "' but the observer is not the tracker in transaction named '" +
+                       getContext().getTransaction().getName() + "'." );
     }
     rawAddObserver( observer );
   }